Fix redis error occurring after integration tests
The integration test was opening a separate redis client connection
which prevented the app's client connection from being closed.

So after the integration tests and the server kill, the open client
would try to reconnect once resulting in one ECONNREFUSED error showing
up after every passing test run.

This removes the separate client from the integration test and the test
now waits for the app to start listening and initialize the client.